# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/ai/codex/images": {
      "post": {
        "operationId": "start_codex_image_generation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodexImageGenerationOptions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodexImageGenerationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          }
        }
      }
    },
    "/config/providers/{id}/secret": {
      "put": {
        "summary": "Save (or overwrite) the keyring secret for a provider. Creates the\nprovider entry in `config.providers` if it didn't exist. `PUT` because\nsetting the secret is idempotent for the same body.",
        "operationId": "set_provider_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      },
      "delete": {
        "summary": "Clear a provider's keyring secret. The provider entry itself is kept.",
        "operationId": "clear_provider_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/downloads": {
      "get": {
        "operationId": "list_downloads",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDownloadsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "start_download",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartDownloadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/google-fonts/{family}/fetch": {
      "post": {
        "operationId": "fetch_google_font",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "description": "Google Fonts family name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/google-fonts/{family}/{file}": {
      "get": {
        "operationId": "get_google_font_file",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "description": "Google Fonts family name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file",
            "in": "path",
            "description": "Font filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "font/ttf": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Raw binary content in the HTTP API."
                }
              }
            }
          }
        }
      }
    },
    "/history/apply": {
      "post": {
        "operationId": "apply_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Op"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/operations/{id}": {
      "delete": {
        "operationId": "cancel_operation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppConfig": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DataConfig"
              }
            ],
            "default": {
              "path": "<app-data>/Yomika"
            }
          },
          "http": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/HttpConfig"
              }
            ],
            "default": {
              "connect_timeout": 20,
              "max_retries": 3,
              "read_timeout": 300
            }
          },
          "pipeline": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            ],
            "default": {
              "bubble_segmenter": "speech-bubble-segmentation",
              "detector": "pp-doclayout-v3",
              "font_detector": "yuzumarker-font-detection",
              "inpainter": "lama-manga",
              "ocr": "paddle-ocr-vl-1.6",
              "renderer": "yomika-renderer",
              "segmenter": "comic-text-detector-seg",
              "translator": "llm"
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderConfig"
            },
            "default": []
          }
        }
      },
...
      "CodexImageGenerationOptions": {
        "type": "object",
        "required": ["pageId", "prompt"],
        "properties": {
          "instructions": {
            "type": ["string", "null"]
          },
          "model": {
            "type": ["string", "null"]
          },
          "pageId": {
            "$ref": "#/components/schemas/PageId"
          },
          "prompt": {
            "type": "string"
          },
          "quality": {
            "type": ["string", "null"]
          },
          "size": {
            "type": ["string", "null"]
          }
        }
      },
      "CodexImageGenerationResponse": {
        "type": "object",
        "required": ["operationId"],
        "properties": {
          "operationId": {
            "type": "string"
          }
        }
      },
      "ConfigPatch": {
        "type": "object",
        "description": "Sparse patch for `yomika_app::AppConfig`. Missing fields mean \"leave\nas-is\". The `providers` field, if present, replaces the whole provider\nlist — we do not merge by id because ordering is meaningful.",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DataConfigPatch"
              }
            ]
          },
          "http": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HttpConfigPatch"
              }
            ]
          },
          "pipeline": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PipelineConfigPatch"
              }
            ]
          },
          "providers": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/ProviderPatch"
            },
            "description": "If present, replaces the entire list. Api_key values of `\"[REDACTED]\"`\nare interpreted as \"leave the existing secret alone\"."
          }
        }
      },
...
      "DataConfig": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
      "DataConfigPatch": {
        "type": "object",
        "properties": {
          "path": {
            "type": ["string", "null"]
          }
        }
      },
...
      "HistoryResult": {
        "type": "object",
        "properties": {
          "epoch": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "New epoch. `None` only for a no-op undo/redo at the stack boundary.",
            "minimum": 0
          }
        }
      },
      "HttpConfig": {
        "type": "object",
        "properties": {
          "connect_timeout": {
            "type": "integer",
            "format": "int64",
            "default": 20,
            "minimum": 0
          },
          "max_retries": {
            "type": "integer",
            "format": "int32",
            "default": 3,
            "minimum": 0
          },
          "read_timeout": {
            "type": "integer",
            "format": "int64",
            "default": 300,
            "minimum": 0
          }
        }
      },
...
      "ListDownloadsResponse": {
        "type": "object",
        "required": ["downloads"],
        "properties": {
          "downloads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DownloadProgress"
            }
          }
        }
      },
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "Op": {
        "oneOf": [
          {
            "type": "object",
            "required": ["updateProjectMeta"],
            "properties": {
              "updateProjectMeta": {
                "type": "object",
                "required": ["patch"],
                "properties": {
                  "patch": {
                    "$ref": "#/components/schemas/ProjectMetaPatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/ProjectMetaPatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["addPage"],
            "properties": {
              "addPage": {
                "type": "object",
                "required": ["page", "at"],
                "properties": {
                  "at": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "page": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["removePage"],
            "properties": {
              "removePage": {
                "type": "object",
                "required": ["id", "prev_page", "prev_index"],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "prev_index": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "prev_page": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["updatePage"],
            "properties": {
              "updatePage": {
                "type": "object",
                "required": ["id", "patch"],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "patch": {
                    "$ref": "#/components/schemas/PagePatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/PagePatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["reorderPages"],
            "properties": {
              "reorderPages": {
                "type": "object",
                "required": ["order", "prev_order"],
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PageId"
                    }
                  },
                  "prev_order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PageId"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["addNode"],
            "properties": {
              "addNode": {
                "type": "object",
                "required": ["page", "node", "at"],
                "properties": {
                  "at": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "node": {
                    "$ref": "#/components/schemas/Node"
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["removeNode"],
            "properties": {
              "removeNode": {
                "type": "object",
                "required": ["page", "id", "prev_node", "prev_index"],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/NodeId"
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "prev_index": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "prev_node": {
                    "$ref": "#/components/schemas/Node"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["updateNode"],
            "properties": {
              "updateNode": {
                "type": "object",
                "required": ["page", "id", "patch"],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/NodeId"
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "patch": {
                    "$ref": "#/components/schemas/NodePatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/NodePatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["reorderNodes"],
            "properties": {
              "reorderNodes": {
                "type": "object",
                "required": ["page", "order", "prev_order"],
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/NodeId"
                    }
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "prev_order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/NodeId"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["batch"],
            "properties": {
              "batch": {
                "type": "object",
                "required": ["ops", "label"],
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "ops": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Op"
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ProviderConfig": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "api_key": {
            "type": ["string", "null"],
            "description": "Populated from credential storage on `load()`, never written to config.toml.\nSerializes as `\"[REDACTED]\"` in API responses."
          },
          "base_url": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "ProviderSecretRequest": {
        "type": "object",
        "required": ["secret"],
        "properties": {
          "secret": {
            "type": "string"
          }
        }
      },
...
      "StartDownloadRequest": {
        "type": "object",
        "required": ["modelId"],
        "properties": {
          "modelId": {
            "type": "string",
            "description": "Package id, as declared via `declare_hf_model_package!`\n(e.g. `\"model:comic-text-detector:yolo-v5\"`)."
          }
        }
      },
      "StartDownloadResponse": {
        "type": "object",
        "required": ["operationId"],
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Operation id. Reusing the package id keeps ids meaningful for clients\nwatching progress events."
          }
        }
      },
...
    }
  }
}